+++ differentiable_robot_model/__init__.py
@@ -4,7 +4,7 @@ Differentiable Robot Model
 Differentiable and learnable robot model.
 """
 
-from .differentiable_robot_model import (
+from .robot_model import (
     DifferentiableRobotModel,
     DifferentiableKUKAiiwa,
     DifferentiableFrankaPanda,
