+++ examples/learn_forward_dynamics_iiwa.py
@@ -11,11 +11,11 @@ from hydra.experimental import compose as hydra_compose
 from hydra.experimental import initialize_config_dir
 
 import differentiable_robot_model
-from differentiable_robot_model.differentiable_robot_model import (
+from differentiable_robot_model.robot_model import (
     DifferentiableRobotModel,
     DifferentiableKUKAiiwa,
 )
-from differentiable_robot_model.data_generation_utils import (
+from differentiable_robot_model.data_utils import (
     generate_sine_motion_forward_dynamics_data,
 )
 import diff_robot_data
