+++ tests/test_devices.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 import diff_robot_data
-from differentiable_robot_model.differentiable_robot_model import (
+from differentiable_robot_model.robot_model import (
     DifferentiableRobotModel,
     LearnableRigidBodyConfig,
 )
